27C

+++ 17-the-wild-oasis/wildo/src/services/apiCabins.js
@@ -1,4 +1,4 @@
-import supabase, {supabaseUrl} from "./supabase";
+import supabase, { supabaseUrl } from "./supabase";
 
 export async function getCabins() {
 
@@ -23,32 +23,33 @@ export async function deleteCabin(id) {
 export async function createEditCabin(newCabin, id) {
     const hasImagePath = newCabin?.image?.startsWith?.(supabaseUrl)
 
-    const imageName = `${Math.random()}-${newCabin?.image?.name}`.replaceAll("/","")
+    const imageName = `${Math.random()}-${newCabin?.image?.name}`.replaceAll("/", "")
 
     const imagePath = hasImagePath ? newCabin.image : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`
     // 1. Create Cabin
     let query = supabase.from('cabins');
 
-    if(!id) query = query.insert([{...newCabin, image: imagePath}])
-   
-    if(id) query = query.update({ ...newCabin, image: imagePath })
-        .eq('id', id)    
-    
+    if (!id) query = query.insert([{ ...newCabin, image: imagePath }])
+
+    if (id) query = query.update({ ...newCabin, image: imagePath })
+        .eq('id', id)
+
     const { data, error } = await query.select().single();
-        
+
     if (error) {
         console.log(error)
         throw new Error(error)
     }
 
     // 2. Upload Image
+    if (hasImagePath) return data;
     const { error: storageError } = await supabase
-    .storage
-    .from('cabin-images')
-    .upload(imageName, newCabin.image)
+        .storage
+        .from('cabin-images')
+        .upload(imageName, newCabin.image)
 
     // 3.Delete cabin if there was an error in uploading file
-    if(storageError) {
+    if (storageError) {
         await supabase.from("cabins").delete().eq("id", data.id)
         console.log(storageError);
         throw new Error("Cabin image could not be uploaded")
